requirements.txt
flask==2.3.3
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.1
pydantic==2.5.2
pydantic-settings==2.1.0
aiohttp==3.8.6
python-multipart==0.0.6
jinja2==3.1.2
bootstrap-flask==2.3.0 
beautifulsoup4==4.12.2
lxml==4.9.5
flask[async]
flask-bootstrap5
aiohttp
beautifulsoup4
pydantic-settings